Refactor utils/pageCursor by separating modules
