+++ src/utils/paginator/index.ts
@@ -1,2 +1,1 @@
-export * from './pageCursor';
 export * from './pageEdge';
