+++ src/utils/paginator/pageCursor/pageCursors.ts
@@ -1,139 +1,5 @@
-import { PrismaClient } from '@prisma/client';
-
-interface Props3<T, K> {
-  page: number,
-  pageInfo: {
-    currentPage: number,
-    size: number,
-    totalPages: number,
-  },
-  model: K,
-  findManyArgs: T,
-}
-
-// Returns an opaque cursor for a page.
-async function pageToCursorObject<FindManyArgs>({
-  page,
-  pageInfo,
-  model,
-  findManyArgs,
-}: Props3<FindManyArgs, typeof model>) {
-  const { currentPage, size, totalPages } = pageInfo;
-  let cursorId: number | string;
-  const prisma = new PrismaClient();
-  const prismaModel = prisma[model.name.toLowerCase()];
-
-  // first
-  if (page === 1) {
-    let findManyArgsForFirst: FindManyArgs;
-    // @ts-ignore
-    if (findManyArgs?.orderBy) {
-      // @ts-ignore
-      const { orderBy } = findManyArgs;
-      findManyArgsForFirst = { ...findManyArgsForFirst, orderBy: { ...orderBy } };
-    }
-    // @ts-ignore
-    if (findManyArgs?.where) {
-      // @ts-ignore
-      const { where } = findManyArgs;
-      findManyArgsForFirst = { ...findManyArgsForFirst, where: { ...where } };
-    }
-    const result = await prismaModel.findMany({
-      ...findManyArgsForFirst,
-      take: 1,
-    });
-    cursorId = result[0].id;
-
-  // last
-  } else if (page === totalPages) {
-    let findManyArgsForLast: FindManyArgs;
-    // @ts-ignore
-    if (findManyArgs?.orderBy) {
-      // @ts-ignore
-      const orderByKey = Object.keys(findManyArgs.orderBy)[0];
-      // @ts-ignore
-      const orderDirection = findManyArgs.orderBy[orderByKey] === 'asc' ? 'desc' : 'asc';
-      findManyArgsForLast = {
-        ...findManyArgsForLast,
-        orderBy: {
-          [orderByKey]: orderDirection,
-        },
-      };
-    } else {
-      findManyArgsForLast = {
-        ...findManyArgsForLast,
-        orderBy: {
-          id: 'desc',
-        },
-      };
-    }
-    // @ts-ignore
-    if (findManyArgs?.where) {
-      // @ts-ignore
-      const { where } = findManyArgs;
-      findManyArgsForLast = { ...findManyArgsForLast, where: { ...where } };
-    }
-    const result = await prismaModel.findMany({
-      ...findManyArgsForLast,
-      take: 1,
-    });
-    cursorId = result[0].id;
-
-  // around & previous
-  } else {
-    const distance = (page - currentPage) * size;
-    const takeSkipArgs = {
-      take: distance < 0 ? -1 : 1,
-      skip: distance < 0 ? distance * -1 : distance,
-    };
-    const result = await prismaModel.findMany({
-      ...findManyArgs,
-      ...takeSkipArgs,
-    });
-    cursorId = result[0].id;
-  }
-
-  return {
-    cursor: Buffer.from('saltysalt'.concat(String(cursorId))).toString('base64'),
-    page,
-    isCurrent: currentPage === page,
-  };
-}
-
-interface Props2<T, K> {
-  start: number,
-  end: number,
-  pageInfo: {
-    currentPage: number,
-    size: number,
-    totalPages: number,
-  },
-  model: K,
-  findManyArgs: T,
-}
-
-// Returns an array of PageCursor objects
-// from start to end (page numbers).
-async function pageCursorsToArray<FindManyArgs>({
-  start,
-  end,
-  pageInfo,
-  model,
-  findManyArgs,
-}: Props2<FindManyArgs, typeof model>) {
-  let page;
-  const cursors = [];
-  for (page = start; page <= end; page++) {
-    const cursorResult = await pageToCursorObject<FindManyArgs>({
-      page,
-      pageInfo,
-      model,
-      findManyArgs,
-    });
-    cursors.push(cursorResult);
-  }
-  return cursors;
-}
+import { pageCursorsToArray } from './cursorArray';
+import { pageToCursorObject } from './cursorObject';
 
 // Returns the total number of pagination results capped to PAGE_NUMBER_CAP.
 export function computeTotalPages(totalCount:number, size: number): number {
@@ -172,7 +38,8 @@ export async function createPageCursors<FindManyArgs>({
 }: Props<FindManyArgs, typeof model>): Promise<pageCursors> {
   // If buttonNum is even, bump it up by 1, and log out a warning.
   if (buttonNum % 2 === 0) {
-    console.log(`Max of ${buttonNum} passed to page cursors, using ${buttonNum + 1}`);
+    // eslint-disable-next-line
+    console.log(`buttonNum of ${buttonNum} passed to page cursors, but using ${buttonNum + 1} instead for the pagination logic`);
     buttonNum = buttonNum + 1;
   }
